+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 * Unreleased
+    * Add `common::CstrPrint` helper class which can be passed into the various
+      `printTo()` methods. The content can be retrieved using the
+      `CstrPrint::getCstr()` method which returns a normal `const char*`
+      c-string.
 * 1.1 (2020-04-25, TZ DB version 2020a)
     * Fix broken links in `README.md`.
     * Fix typos in `USER_GUIDE.md` and update its version number to 1.0 as it